refactor: split large django modules
