+++ oregoninvasiveshotline/reports/models/__init__.py
@@ -1,0 +1,4 @@
+from .invite import Invite
+from .report import Report, receiver__generate_icon
+
+__all__ = ["Invite", "Report", "receiver__generate_icon"]
