+++ oregoninvasiveshotline/users/views/editing.py
@@ -1,0 +1,69 @@
+from django.contrib import messages
+from django.shortcuts import get_object_or_404, redirect, render
+
+from oregoninvasiveshotline.users.forms import UserForm
+from oregoninvasiveshotline.users.models import User
+from oregoninvasiveshotline.users.perms import can_list_users, permissions
+from oregoninvasiveshotline.utils.db import will_be_deleted_with
+
+
+@permissions.can_delete_user
+def delete(request, user_id):
+    user = get_object_or_404(User, pk=user_id)
+    if request.method == "POST":
+        user.delete()
+        messages.success(request, "User deleted!")
+        return redirect("users-list")
+
+    related_objects = list(will_be_deleted_with(user))
+
+    return render(request, "delete.html", {
+        # we don't call this template variable "user" because that collides
+        # with the "user" variable which references the currently logged in
+        # user
+        "object": user,
+        "will_be_deleted_with": related_objects,
+    })
+
+
+@permissions.can_create_user
+def create(request):
+    """
+    Create a new user
+    """
+    return _edit(request)
+
+
+@permissions.can_edit_user
+def edit(request, user_id):
+    """
+    Edit an existing user.
+    """
+    user = get_object_or_404(User, pk=user_id)
+    return _edit(request, user)
+
+
+def _edit(request, user=None):
+    """
+    Handle creating or editing a user
+    """
+    if request.POST:
+        form = UserForm(request.POST, request.FILES, user=request.user, instance=user)
+
+        if form.is_valid():
+            is_new = user is None or user.pk is None
+            user = form.save()
+            if is_new:
+                messages.success(request, "Created!")
+            else:
+                messages.success(request, "Edited!")
+
+            if can_list_users(request.user):
+                return redirect("users-list")
+            return redirect("users-home")
+    else:
+        form = UserForm(user=request.user, instance=user)
+
+    return render(request, 'users/edit.html', {
+        "form": form,
+    })
